dbus_objects: put the object submodule on the top level module
